# Conflicts:
#	GameEngine/GameLogic/Scene/Scene.cpp

+++ GameEngine/ResourceEngine/ResourceManager/BRFImporterHandler/BRFImporterHandler.cpp
@@ -173,7 +173,8 @@ void BRFImporterHandler::LoadFile(std::string fileName, bool mesh, bool material
 		tempMaterial.materialID = materialID;
 		materialID++;
 
-		for (unsigned int j = meshes->size()-meshsize; j < meshes->size(); j++)
+
+		for (size_t j = meshes->size()-meshsize; j < meshes->size(); j++)
 		{
 			unsigned int importedMatID = meshes->at(j).GetMaterialID();
 			if (importedMatID == tempMaterialID)
